add null url for product not pic

+++ app/Product.php
@@ -7,16 +7,16 @@ use Illuminate\Database\Eloquent\Model;
 class Product extends Model
 {
     //
-    protected $guarded =[];
-    protected $appends =['image_url'];
+    protected $guarded = [];
+    protected $appends = ['image_url'];
 
     public function storeProduct($pic)
     {
-        if (!empty(request()->file('image')))
-        {
+        if (!empty(request()->file('image'))) {
             $get_full_name_pic = $pic->getClientOriginalName();
-            $get_path_pic = $pic->storeAs('upload', $get_full_name_pic,'asset');
+            $get_path_pic = $pic->storeAs('upload', $get_full_name_pic, 'asset');
             $ProductPic = pathinfo($get_path_pic, PATHINFO_BASENAME);
+
             $this->update([
                 'image' => $ProductPic
             ]);
@@ -25,27 +25,53 @@ class Product extends Model
 
     public function updateProduct($pic)
     {
-        if (!empty(request()->file('image')))
+        if(!empty(request()->file('image')))
         {
-            unlink(public_path('picture/upload/'.$this->image));
-            $get_full_name_pic = $pic->getClientOriginalName();
-            $get_path_pic = $pic->storeAs('upload', $get_full_name_pic,'asset');
-            $ProductPic = pathinfo($get_path_pic, PATHINFO_BASENAME);
-            $this->update([
-                'image' => $ProductPic
-            ]);
+            if(!empty($this->image)){
+                unlink('picture/upload/' . $this->image);
+
+                $get_full_name_pic = $pic->getClientOriginalName();
+                $get_path_pic = $pic->storeAs('upload', $get_full_name_pic, 'asset');
+                $ProductPic = pathinfo($get_path_pic, PATHINFO_BASENAME);
+
+                $this->update([
+                    'image' => $ProductPic
+                ]);
+            }
+            else{
+                $get_full_name_pic = $pic->getClientOriginalName();
+                $get_path_pic = $pic->storeAs('upload', $get_full_name_pic, 'asset');
+                $ProductPic = pathinfo($get_path_pic, PATHINFO_BASENAME);
+
+                $this->update([
+                    'image' => $ProductPic
+                ]);
+            }
         }
     }
 
     public function deleteProduct()
     {
-        unlink('picture/upload/'.$this->image);
-        $this->delete();
+        if ($this->image != null) {
+            unlink('picture/upload/' . $this->image);
+
+            $this->delete();
+
+        } else {
+            $this->delete();
+        }
+
     }
 
     public function getImageUrlAttribute()
     {
-        $url =  URL('')."/api/v1/download/$this->image";
-        return $url;
+        if(!empty($this->image)){
+            $url = URL('') . "/api/v1/download/$this->image";
+            return $url;
+        }else
+        {
+            return null;
+        }
+
     }
 }
